requirements.txt
Pillow==3.4.1
pytesseract==0.1.6
